Support Python 3.3 and 3.4
TL;DR: kMer supports Python 2 and 3 and every module has the following
line at the top:

    from __future__ import (absolute_import, division, print_function,
                            unicode_literals)

We now support Python versions 2.6, 2.7, 3.3, and 3.4 in a single
codebase without using 2to3. We don't support Python 3.2 because
BioPython does not.

We use the Python future package [1] as a compatibility layer between
Python 2 and Python 3. The goal is to use a single, clean Python 3.x-
compatible codebase to support both Python 2 and Python 3 with minimal
overhead.

Most changes are quite straightforward (e.g., absolute imports, print
statement, division operator). The main painpoint is of course the
bytestring versus unicode story. We now import `unicode_literals` [2] in
each module and maintain that all text in kMer is unicode (`unicode` in
Python 2, `str` in Python 3).

[1] http://python-future.org/
[2] http://python-future.org/imports.html#should-i-import-unicode-literals