Disallow profile names containing / or .
The / and . characters are not allowed in HDF5 group or dataset
names. For now we are aborting with an error message if trying to
use a profile name containing / or ..

A more user-friendly approach might be to silently replace the
offending characters? Either one-way by some other character like
| and , or two-way by some magic string like `@@FWDSLASH@@`.

Alternatively, we could store the profile names in a separate
table instead of using them for the dataset names.